+++ .github/workflows/actions.yml
@@ -98,4 +98,3 @@ jobs:
     env:
       SURGE_LOGIN: ${{ secrets.SURGE_LOGIN }}
       SURGE_TOKEN: ${{ secrets.SURGE_TOKEN }}
-      REPO_TOKEN: ${{ secrets.GITHUB_TOKEN }}
