+++ package.json
@@ -19,8 +19,8 @@
     "watch": "lerna run --parallel --ignore ${LERNA_IGNORE:-'wix-rich-content-@(test-env|example|storybook)'} watch",
     "watch:core": "LERNA_IGNORE='wix-rich-content-@(plugin-*|fullscreen)' npm run watch",
     "release": "node scripts/publishPackages",
-    "deploy": "node scripts/deployExamples",
-    "deployPRComment": "node scripts/deployPRComment",
+    "deploy": "node scripts/deployment/deployExamples",
+    "deployPRComment": "node scripts/deployment/deployPRComment",
     "createVersion": "node scripts/createVersion",
     "storybook": "npm run start --prefix examples/storybook",
     "exampleApp": "npm run start --prefix examples/main",
